[VCDA-3721] Support ranges not identical to gateway ranges (#82)
Earlier we would try to match one of the subnets in the gateway in order to
reduce the search space. Now that part of code has been rewritten to just
check if an 'unused ip' in the user-specified subnet also lies in the list of
gateway ranges specified. This approach yields a performant search
method.